Move 3.8 files into base so it's not the whole codebase that is split
