+++ gTile@shuairan/src/3_8/tsconfig.json
@@ -1,6 +1,7 @@
 {
 	"extends": "../../tsconfig.json",
 	"include": [
-		"./**/*"
+		"./**/*",
+		"../base/**/*"
 	],
 }
