+++ gTile@shuairan/src/5_4/tsconfig.json
@@ -2,5 +2,6 @@
 	"extends": "../../tsconfig.json",
 	"include": [
 		"./**/*"
+		"../base/**/*"
 	],
 }
