+++ gTile@shuairan/src/base/ui/TopBar.ts
@@ -1,4 +1,4 @@
-import { App } from "../extension";
+import { App } from "../app";
 
 const St = imports.gi.St;
